requirements.txt
google-generativeai
yfinance
pandas
numpy
prophet